Rename ai_fuctions file
